Refactor TipCalculatorPage by moving it to its own folder and updating import in App
